ames dataloader (#15)

+++ acquistion_functions/optimize.py
@@ -38,6 +38,6 @@ def optimize(*,
     else:
         raise ValueError(f"Invalid mode: {mode}. Must be either 'max' or 'min'.")
     
-    indices = best_rows[:, 1]
+    indices = best_rows[:, 1].astype(np.int64) # since the row was a float dtype (cause we concatenated with the values_with_indices), we need to convert the indices back to int
     num_candidates = min(max_num_results, len(indices)) # guards against the case where there are fewer than max_num_results candidates
     return np.random.choice(indices, num_candidates, replace=False)
